+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Internal] `SegmentTrackingFeature` to always track the latest `n` downloaded elements
 - [Internal] `ErrorDetailsFeature` to track additional details in case an error occurs
 - [Internal] Added code formatting
+- Ability to attach the `ExoPlayerCollector` to a player that's already in a loading state.
 
 ## v1.22.0-beta